#131 [BE] 학교 도메인 분리 및 테스트 코드 작성 (#132)
* chore: school 앱 생성 및 초기화

* chore: account 마이그레이션 초기화

* chore: school fixture 수정

* chore: school 앱 전역 등록

* refactor: Account 앱으로 부터 School 도메인 분리

* refactor: create_user의 College, Department 주소 변경

* test: school 앱 테스트 코드 작성
