+++ backend/fixtures/department_10.json
@@ -1,6 +1,6 @@
 [
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 76,
     "fields": {
       "department_name": "음악학과",
@@ -8,7 +8,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 77,
     "fields": {
       "department_name": "한국음악학과",
@@ -16,7 +16,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 78,
     "fields": {
       "department_name": "미술학과",
@@ -24,7 +24,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 79,
     "fields": {
       "department_name": "조형학과",
@@ -32,7 +32,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 80,
     "fields": {
       "department_name": "디자인학과",
@@ -40,7 +40,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 81,
     "fields": {
       "department_name": "무용학과",
@@ -48,7 +48,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 82,
     "fields": {
       "department_name": "예술문화영상학과",
