+++ backend/fixtures/department_11.json
@@ -1,6 +1,6 @@
 [
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 83,
     "fields": {
       "department_name": "나노메카트로닉스공학과",
@@ -8,7 +8,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 84,
     "fields": {
       "department_name": "나노에너지공학과",
@@ -16,7 +16,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 85,
     "fields": {
       "department_name": "광메카트로닉스공학과",
