+++ backend/fixtures/department_12.json
@@ -1,6 +1,6 @@
 [
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 86,
     "fields": {
       "department_name": "식물생명과학과",
@@ -8,7 +8,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 87,
     "fields": {
       "department_name": "원예생명과학과",
@@ -16,7 +16,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 88,
     "fields": {
       "department_name": "동물생명자원과학과",
@@ -24,7 +24,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 89,
     "fields": {
       "department_name": "식품공학과",
@@ -32,7 +32,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 90,
     "fields": {
       "department_name": "생명환경화학과",
@@ -40,7 +40,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 91,
     "fields": {
       "department_name": "바이오소재과학과",
@@ -48,7 +48,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 92,
     "fields": {
       "department_name": "바이오산업기계공학과",
@@ -56,7 +56,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 93,
     "fields": {
       "department_name": "조경학과",
@@ -64,7 +64,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 94,
     "fields": {
       "department_name": "식품자원경제학과",
@@ -72,7 +72,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 95,
     "fields": {
       "department_name": "IT응용공학과",
@@ -80,7 +80,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 96,
     "fields": {
       "department_name": "바이오환경에너지학과",
