+++ backend/fixtures/department_4.json
@@ -1,6 +1,6 @@
 [
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 30,
     "fields": {
       "department_name": "기계공학부",
@@ -8,7 +8,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 31,
     "fields": {
       "department_name": "고분자공학과",
@@ -16,7 +16,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 32,
     "fields": {
       "department_name": "유기소재시스템공학과",
@@ -24,7 +24,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 33,
     "fields": {
       "department_name": "화공생명·환경공학부-화공생명공학전공",
@@ -32,7 +32,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 34,
     "fields": {
       "department_name": "화공생명·환경공학부-환경공학전공",
@@ -40,7 +40,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 35,
     "fields": {
       "department_name": "전기전자공학부-전자공학전공",
@@ -48,7 +48,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 36,
     "fields": {
       "department_name": "전기전자공학부-전기공학전공",
@@ -56,7 +56,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 37,
     "fields": {
       "department_name": "전기전자공학부-반도체공학전공",
@@ -64,7 +64,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 38,
     "fields": {
       "department_name": "조선해양공학과",
@@ -72,7 +72,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 39,
     "fields": {
       "department_name": "재료공학부",
@@ -80,7 +80,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 40,
     "fields": {
       "department_name": "산업공학과",
@@ -88,7 +88,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 41,
     "fields": {
       "department_name": "항공우주공학과",
@@ -96,7 +96,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 42,
     "fields": {
       "department_name": "건축공학과",
@@ -104,7 +104,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 43,
     "fields": {
       "department_name": "건축학과",
@@ -112,7 +112,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 44,
     "fields": {
       "department_name": "도시공학과",
@@ -120,7 +120,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 45,
     "fields": {
       "department_name": "사회기반시스템공학과",
