+++ backend/fixtures/department_6.json
@@ -1,6 +1,6 @@
 [
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 63,
     "fields": {
       "department_name": "무역학부",
@@ -8,7 +8,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 64,
     "fields": {
       "department_name": "경제학부",
@@ -16,7 +16,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 65,
     "fields": {
       "department_name": "관광컨벤션학과",
@@ -24,7 +24,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 66,
     "fields": {
       "department_name": "국제학부",
@@ -32,7 +32,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 67,
     "fields": {
       "department_name": "공공정책학부",
