+++ backend/fixtures/department_7.json
@@ -1,6 +1,6 @@
 [
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 68,
     "fields": {
       "department_name": "경영학과",
