+++ backend/fixtures/department_9.json
@@ -1,6 +1,6 @@
 [
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 71,
     "fields": {
       "department_name": "아동가족학과",
@@ -8,7 +8,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 72,
     "fields": {
       "department_name": "의류학과",
@@ -16,7 +16,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 73,
     "fields": {
       "department_name": "식품영양학과",
@@ -24,7 +24,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 74,
     "fields": {
       "department_name": "실내환경디자인학과",
@@ -32,7 +32,7 @@
     }
   },
   {
-    "model": "account.department",
+    "model": "school.department",
     "pk": 75,
     "fields": {
       "department_name": "스포츠과학과",
